Inline AppConsentDetailsComponent into AppConsentComponent
Move the specs into appropriate places
